deleted mocks, api moved in separated directories
