+++ src/pages/admin/team/components/team-page-content/TeamPageContent.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react';
 import { StatusFilter } from '../../../../../types/Common';
 import { MembersList } from '../members-list/MembersList';
 import { MemberFormValues } from '../member-form/MemberForm';
-import { TeamMembersApi } from '../../../../../services/data-fetch/admin-page-data-fetch/team-page-data-fetch/TeamMembersApi';
+import { TeamMembersApi } from '../../../../../services/data-fetch/admin-page-data-fetch/team-page-data-fetch/TeamMembersApi/TeamMembersApi';
 import { useAdminClient } from '../../../../../utils/hooks/use-admin-client/useAdminClient';
 import { TeamPageToolbar } from '../team-page-toolbar/TeamPageToolbar';
