--- src/utils/mock-data/admin-page/teamPage.ts
@@ -1,664 +1,0 @@
-import { TeamMember } from '../../../types/admin/TeamMembers';
-
-export const mockMembers: TeamMember[] = [
-    {
-        id: 1,
-        img: 'https://randomuser.me/api/portraits/women/1.jpg',
-        fullName: 'John Brown',
-        description: 'Software Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 2,
-        img: 'https://randomuser.me/api/portraits/men/2.jpg',
-        fullName: 'John Jones',
-        description: 'HR Specialist',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Основна команда',
-            description: 'Test',
-        },
-    },
-    {
-        id: 3,
-        img: 'https://randomuser.me/api/portraits/women/3.jpg',
-        fullName: 'Emma Anderson',
-        description: 'DevOps Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 4,
-        img: 'https://randomuser.me/api/portraits/men/4.jpg',
-        fullName: 'Daniel Miller',
-        description: 'HR Specialist',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 5,
-        img: 'https://randomuser.me/api/portraits/women/5.jpg',
-        fullName: 'John Williams',
-        description: 'Software Engineer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 6,
-        img: 'https://randomuser.me/api/portraits/men/6.jpg',
-        fullName: 'Mia Garcia',
-        description: 'Marketing Manager',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 7,
-        img: 'https://randomuser.me/api/portraits/women/7.jpg',
-        fullName: 'Mia Brown',
-        description: 'Marketing Manager',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 8,
-        img: 'https://randomuser.me/api/portraits/men/8.jpg',
-        fullName: 'Michael Williams',
-        description: 'DevOps Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 9,
-        img: 'https://randomuser.me/api/portraits/women/9.jpg',
-        fullName: 'Olivia Davis',
-        description: 'Frontend Developer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 10,
-        img: 'https://randomuser.me/api/portraits/men/10.jpg',
-        fullName: 'John Brown',
-        description: 'QA Analyst',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 11,
-        img: 'https://randomuser.me/api/portraits/women/11.jpg',
-        fullName: 'Olivia Garcia',
-        description: 'DevOps Engineer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 12,
-        img: 'https://randomuser.me/api/portraits/men/12.jpg',
-        fullName: 'Daniel Smith',
-        description: 'Business Analyst',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 13,
-        img: 'https://randomuser.me/api/portraits/women/13.jpg',
-        fullName: 'David Jones',
-        description: 'Marketing Manager',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 14,
-        img: 'https://randomuser.me/api/portraits/men/14.jpg',
-        fullName: 'David Miller',
-        description: 'DevOps Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 15,
-        img: 'https://randomuser.me/api/portraits/women/15.jpg',
-        fullName: 'Olivia Garcia',
-        description: 'Backend Developer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 16,
-        img: 'https://randomuser.me/api/portraits/men/16.jpg',
-        fullName: 'David Garcia',
-        description: 'Marketing Manager',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 17,
-        img: 'https://randomuser.me/api/portraits/women/17.jpg',
-        fullName: 'Mia Garcia',
-        description: 'HR Specialist',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 18,
-        img: 'https://randomuser.me/api/portraits/men/18.jpg',
-        fullName: 'Michael Johnson',
-        description: 'Frontend Developer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 19,
-        img: 'https://randomuser.me/api/portraits/women/19.jpg',
-        fullName: 'Mia Johnson',
-        description: 'Project Manager',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 20,
-        img: 'https://randomuser.me/api/portraits/men/20.jpg',
-        fullName: 'Mia Smith',
-        description: 'DevOps Engineer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 21,
-        img: 'https://randomuser.me/api/portraits/women/21.jpg',
-        fullName: 'James Williams',
-        description: 'Frontend Developer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 22,
-        img: 'https://randomuser.me/api/portraits/men/22.jpg',
-        fullName: 'David Garcia',
-        description: 'Frontend Developer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 23,
-        img: 'https://randomuser.me/api/portraits/women/23.jpg',
-        fullName: 'James Davis',
-        description: 'QA Analyst',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 24,
-        img: 'https://randomuser.me/api/portraits/men/24.jpg',
-        fullName: 'Daniel Miller',
-        description: 'Backend Developer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 25,
-        img: 'https://randomuser.me/api/portraits/women/25.jpg',
-        fullName: 'Daniel Smith',
-        description: 'Marketing Manager',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 26,
-        img: 'https://randomuser.me/api/portraits/men/26.jpg',
-        fullName: 'Daniel Brown',
-        description: 'DevOps Engineer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 27,
-        img: 'https://randomuser.me/api/portraits/women/27.jpg',
-        fullName: 'John Jones',
-        description: 'Business Analyst',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 28,
-        img: 'https://randomuser.me/api/portraits/men/28.jpg',
-        fullName: 'John Smith',
-        description: 'UI/UX Designer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 29,
-        img: 'https://randomuser.me/api/portraits/women/29.jpg',
-        fullName: 'Emma Williams',
-        description: 'DevOps Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 30,
-        img: 'https://randomuser.me/api/portraits/men/30.jpg',
-        fullName: 'Michael Miller',
-        description: 'Software Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 31,
-        img: 'https://randomuser.me/api/portraits/women/31.jpg',
-        fullName: 'Ava Johnson',
-        description: 'Backend Developer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 32,
-        img: 'https://randomuser.me/api/portraits/men/32.jpg',
-        fullName: 'Olivia Anderson',
-        description: 'Project Manager',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 33,
-        img: 'https://randomuser.me/api/portraits/women/33.jpg',
-        fullName: 'David Johnson',
-        description: 'Marketing Manager',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 34,
-        img: 'https://randomuser.me/api/portraits/men/34.jpg',
-        fullName: 'Daniel Brown',
-        description: 'UI/UX Designer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 35,
-        img: 'https://randomuser.me/api/portraits/women/35.jpg',
-        fullName: 'David Johnson',
-        description: 'Marketing Manager',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 36,
-        img: 'https://randomuser.me/api/portraits/men/36.jpg',
-        fullName: 'Ava Anderson',
-        description: 'Marketing Manager',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 37,
-        img: 'https://randomuser.me/api/portraits/women/37.jpg',
-        fullName: 'John Brown',
-        description: 'Project Manager',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 38,
-        img: 'https://randomuser.me/api/portraits/men/38.jpg',
-        fullName: 'James Davis',
-        description: 'Project Manager',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 39,
-        img: 'https://randomuser.me/api/portraits/women/39.jpg',
-        fullName: 'Emma Smith',
-        description: 'UI/UX Designer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 40,
-        img: 'https://randomuser.me/api/portraits/men/40.jpg',
-        fullName: 'Mia Brown',
-        description: 'Software Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 41,
-        img: 'https://randomuser.me/api/portraits/women/41.jpg',
-        fullName: 'Olivia Williams',
-        description: 'Backend Developer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 42,
-        img: 'https://randomuser.me/api/portraits/men/42.jpg',
-        fullName: 'David Jones',
-        description: 'DevOps Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 43,
-        img: 'https://randomuser.me/api/portraits/women/43.jpg',
-        fullName: 'Mia Wilson',
-        description: 'QA Analyst',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 44,
-        img: 'https://randomuser.me/api/portraits/men/44.jpg',
-        fullName: 'Mia Brown',
-        description: 'Marketing Manager',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 45,
-        img: 'https://randomuser.me/api/portraits/women/45.jpg',
-        fullName: 'Michael Brown',
-        description: 'UI/UX Designer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 46,
-        img: 'https://randomuser.me/api/portraits/men/46.jpg',
-        fullName: 'James Davis',
-        description: 'HR Specialist',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 47,
-        img: 'https://randomuser.me/api/portraits/women/47.jpg',
-        fullName: 'James Brown',
-        description: 'Marketing Manager',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 48,
-        img: 'https://randomuser.me/api/portraits/men/48.jpg',
-        fullName: 'Olivia Brown',
-        description: 'HR Specialist',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 49,
-        img: 'https://randomuser.me/api/portraits/women/49.jpg',
-        fullName: 'Olivia Miller',
-        description: 'DevOps Engineer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-    {
-        id: 50,
-        img: 'https://randomuser.me/api/portraits/men/50.jpg',
-        fullName: 'James Jones',
-        description: 'Business Analyst',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 51,
-        img: 'https://randomuser.me/api/portraits/women/51.jpg',
-        fullName: 'John Jones',
-        description: 'DevOps Engineer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 52,
-        img: 'https://randomuser.me/api/portraits/men/52.jpg',
-        fullName: 'Mia Williams',
-        description: 'DevOps Engineer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 53,
-        img: 'https://randomuser.me/api/portraits/women/53.jpg',
-        fullName: 'John Anderson',
-        description: 'Software Engineer',
-        status: 'Опубліковано',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 54,
-        img: 'https://randomuser.me/api/portraits/men/54.jpg',
-        fullName: 'James Miller',
-        description: 'Backend Developer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Наглядова рада',
-            description: 'Test',
-        },
-    },
-    {
-        id: 55,
-        img: 'https://randomuser.me/api/portraits/women/55.jpg',
-        fullName: 'Emma Wilson',
-        description: 'DevOps Engineer',
-        status: 'Чернетка',
-        category: {
-            id: 1,
-            name: 'Радники',
-            description: 'Test',
-        },
-    },
-];
